+ Fix a bug in calculator/getop.cpp on reading input like 0.123.123 + Fix a bug in ch4/4-2.cpp that uses int to store double value + Fix a bug in ch4/4-11.cpp on reading single character '.' + Include common.mk in calculator/Makefile

+++ ch4/4-2.cpp
@@ -40,7 +40,7 @@ double atof(const char s[])
         i++;
     }
 
-    int val = 0;
+    double val = 0;
     for (; isdigit(s[i]); i++) {
         val = 10 * val + (s[i] - '0');
     }
